# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sso/token": {
            "post": {
                "tags": [
                    "SSO"
                ],
                "summary": "Issue access token untuk aplikasi client.",
                "description": "Login employee dan dapatkan JWT access token untuk SSO",
                "operationId": "ssoToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "client_secret",
                                    "email",
                                    "password",
                                    "grant_type"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "string"
                                    },
                                    "client_secret": {
                                        "type": "string"
                                    },
                                    "grant_type": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token berhasil diterbitkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/sso/userinfo": {
            "get": {
                "tags": [
                    "SSO"
                ],
                "summary": "Ambil data user yang sedang login.",
                "description": "Ambil data employee beserta jabatan dan foto profile berdasarkan Bearer token",
                "operationId": "ssoUserinfo",
                "responses": {
                    "200": {
                        "description": "Data user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/sso/introspect": {
            "post": {
                "tags": [
                    "SSO"
                ],
                "summary": "Introspect token.",
                "description": "Validasi access token dan kembalikan klaimnya",
                "operationId": "ssoIntrospect",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "active": {
                                            "type": "boolean"
                                        },
                                        "sub": {
                                            "type": "integer"
                                        },
                                        "exp": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Employee": {
                "description": "Model Employee",
                "properties": {
                    "emp_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "nik_telin": {
                        "type": "string",
                        "nullable": true
                    },
                    "nik_tg": {
                        "type": "string",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "nullable": true
                    },
                    "religion": {
                        "type": "string",
                        "nullable": true
                    },
                    "place_of_birth": {
                        "type": "string",
                        "nullable": true
                    },
                    "date_of_birth": {
                        "type": "string",
                        "nullable": true
                    },
                    "employee_status": {
                        "type": "string",
                        "nullable": true
                    },
                    "position_band": {
                        "type": "string",
                        "nullable": true
                    },
                    "work_location": {
                        "type": "string",
                        "nullable": true
                    },
                    "office_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    },
                    "province": {
                        "type": "string",
                        "nullable": true
                    },
                    "pos_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "telepon": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_city": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_province": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_country": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_pos_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_tlp_no": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "email_asli": {
                        "type": "string",
                        "nullable": true
                    },
                    "profile_photo": {
                        "type": "string",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_status_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "auth_type_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "auth_dtm": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "nonce": {
                        "type": "string",
                        "nullable": true
                    },
                    "token": {
                        "type": "string",
                        "nullable": true
                    },
                    "attempts": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status_login": {
                        "type": "string",
                        "nullable": true
                    },
                    "positions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Position"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Position": {
                "description": "Model Position",
                "properties": {
                    "pos_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "department": {
                        "type": "object",
                        "nullable": true
                    },
                    "pivot": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}